+++ core/src/main/java/io/github/bekoenig/getdown/util/HostWhitelist.java
@@ -6,24 +6,44 @@
 package io.github.bekoenig.getdown.util;
 
 import io.github.bekoenig.getdown.data.Build;
+import io.github.bekoenig.getdown.data.SysProps;
 
 import java.net.MalformedURLException;
 import java.net.URL;
+import java.util.Arrays;
 import java.util.List;
 
 /**
- * Optional support for compiling a URL host whitelist into the Getdown JAR.
+ * Optional support for compiling a URL host whitelist into the Getdown JAR or defined by system
+ * property.
  * Useful if you're on the paranoid end of the security spectrum.
  *
  * @see Build#hostWhitelist()
+ * @see SysProps#hostWhitelist()
  */
 public final class HostWhitelist {
+
+    /**
+     * Returns the active host whitelist. If no host whitelist was set in the build (default
+     * behavior), the system property with the same name is used.
+     */
+    private static List<String> activeHostWhitelist() {
+        // The compiled host whitelist has the highest priority.
+        String hostWhitelist = Build.hostWhitelist();
+        // If the compiled host whitelist is blank, the system property can be used.
+        if (StringUtil.isBlank(hostWhitelist)) {
+            hostWhitelist = SysProps.hostWhitelist();
+        }
+
+        return Arrays.asList(StringUtil.parseStringArray(hostWhitelist));
+    }
+
     /**
      * Verifies that the specified URL should be accessible, per the built-in host whitelist.
      * See {@link Build#hostWhitelist()} and {@link #verify(List, URL)}.
      */
     public static URL verify(URL url) throws MalformedURLException {
-        return verify(Build.hostWhitelist(), url);
+        return verify(activeHostWhitelist(), url);
     }
 
     /**